Moved Class Middleware and Wrapper functions to middleware.ts and wrapper.ts respectively
